变更格式化代码位置
